pyproject.toml
[tool.poetry]
name = "jarvis"
version = "0.1.0"
description = "A callgraph generator for Python"
authors = ["Mark Bekooy <mark.bekooy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
